+++ example/lazy-module.ts
@@ -4,11 +4,11 @@ const plugin = (app: Elysia) => app.get('/plugin', () => 'Plugin')
 const asyncPlugin = async (app: Elysia) => app.get('/async', () => 'A')
 
 const app = new Elysia()
-	// .decorate('a', () => 'hello')
+	.decorate('a', () => 'hello')
 	.use(plugin)
 	.use(import('./lazy'))
 	.use((app) => app.get('/inline', ({ store: { a } }) => 'inline'))
-	// .get('/', ({ a }) => a())
+	.get('/', ({ a }) => a())
 	.listen(3000)
 
 await app.modules